created commentController and modified accordingly
